backend/requirements.txt
Django==5.0.6
psycopg[binary]==3.2.1
djangorestframework==3.16.1
django-filter==24.2
django-cors-headers==4.4.0
drf-writable-nested==0.7.2
drf-extra-fields==3.7.0
django-environ==0.12.0
drf-yasg
dj-database-url
psycopg2-binary